some cleaning.
