+++ .idea/misc.xml
@@ -3,5 +3,5 @@
   <component name="Black">
     <option name="sdkName" value="Python 3.8 (3d_SceneScript)" />
   </component>
-  <component name="ProjectRootManager" version="2" project-jdk-name="3d_scenescript" project-jdk-type="Python SDK" />
+  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.8 (3d_SceneScript)" project-jdk-type="Python SDK" />
 </project>
