+++ scripts/model_3.py
@@ -1,10 +1,13 @@
 import torch
-from tests.point_cloud_sinusoidal_pt_cloud_no_vit import PointCloudTransformerLayer
+from three_d_scene_script.point_cloud_sinusoidal_pt_cloud_no_vit import PointCloudTransformerLayer
 import torch.nn as nn
 import torch.nn.functional as F
 import math
 from enum import Enum
+from three_d_scene_script.gt_processor import SceneScriptProcessor
+import matplotlib.pyplot as plt
 
+# Initialize the models and other components
 encoder_model = PointCloudTransformerLayer().cuda()
 pt_cloud_path = "/home/mseleem/Desktop/3d_model_pt/0/semidense_points.csv.gz"
 points, dist_std = encoder_model.read_points_file(pt_cloud_path)
@@ -34,22 +37,19 @@ class Commands(Enum):
 class TransformerOutputLayer(nn.Module):
     def __init__(self, transformer_dim):
         super(TransformerOutputLayer, self).__init__()
-        self.command_layer = nn.Linear(transformer_dim, 5)  
-        self.param_layer = nn.Linear(transformer_dim, 6)  # Output 6 params for wall, doors, windows 
+        self.command_layer = nn.Linear(transformer_dim, 5)
+        self.param_layer = nn.Linear(transformer_dim, 6)
 
     def forward(self, x):
-        command_logits = self.command_layer(x)  # Shape: (batch_size, sequence_length, 5)
-
-        # Mask the "START" command logits to be very negative
-        command_logits[..., 0] = -float('inf') 
-
-        command_probs = F.softmax(command_logits, dim=-1)  # Shape: (batch_size, sequence_length, 5)
-        parameter_logits = self.param_layer(x)  # Shape: (batch_size, sequence_length, 6)
-        parameters_probs = torch.tanh(parameter_logits)  # Shape: (batch_size, sequence_length, 6) # tanh ko kese lao gi wapis raw logits mein ispe socho
+        command_logits = self.command_layer(x)
+        command_logits[..., 0] = -float('inf')
+        command_probs = F.softmax(command_logits, dim=-1)
+        parameter_logits = self.param_layer(x)
+        parameters_probs = torch.tanh(parameter_logits)
         return command_probs, parameters_probs
 
 def select_parameters(command_probs, parameters_probs):
-    command_indx = command_probs[0, 0].argmax(dim=-1).item() + 1 
+    command_indx = command_probs[0, 0].argmax(dim=-1).item() + 1
     if command_indx == Commands.STOP.value:
         parameters = torch.zeros(6).cuda()
     elif command_indx in [Commands.MAKE_WALL.value, Commands.MAKE_DOOR.value, Commands.MAKE_WINDOW.value]:
@@ -86,23 +86,24 @@ class CrossAttention(nn.Module):
         self.W_query = nn.Linear(d_model, d_out_kq)
         self.W_key = nn.Linear(d_model, d_out_kq)
         self.W_value = nn.Linear(d_model, d_out_v)
-    
-    def forward(self, x_1, x_2):
+
+    def forward(self, x_1, x_2, mask=None):
         queries_1 = self.W_query(x_1)
         keys_2 = self.W_key(x_2)
         values_2 = self.W_value(x_2)
-        
+
         attn_scores = queries_1.matmul(keys_2.transpose(-2, -1))
+        if mask is not None:
+            attn_scores = attn_scores.masked_fill(mask == 0, float('-inf'))
         attn_weights = torch.softmax(attn_scores / self.d_out_kq**0.5, dim=-1)
-        
+
         context_vec = attn_weights.matmul(values_2)
         return context_vec
-
 
 class CustomTransformerDecoderLayer(nn.Module):
     def __init__(self, d_model, d_out_kq, d_out_v, dim_feedforward=2048, dropout=0.1):
         super(CustomTransformerDecoderLayer, self).__init__()
-        self.self_attn = nn.MultiheadAttention(d_model, 1, dropout=dropout) # Self-attention: tgt, tgt, tgt
+        self.self_attn = nn.MultiheadAttention(d_model, 1, dropout=dropout)
         self.cross_attention = CrossAttention(d_model, d_out_kq, d_out_v)
         self.linear1 = nn.Linear(d_model, dim_feedforward)
         self.dropout = nn.Dropout(dropout)
@@ -115,17 +116,14 @@ class CustomTransformerDecoderLayer(nn.Module):
         self.dropout3 = nn.Dropout(dropout)
 
     def forward(self, tgt, memory, tgt_mask=None, memory_mask=None, tgt_key_padding_mask=None, memory_key_padding_mask=None):
-        # Self-attention
         tgt2, _ = self.self_attn(tgt, tgt, tgt, attn_mask=tgt_mask, key_padding_mask=tgt_key_padding_mask)
         tgt = tgt + self.dropout1(tgt2)
         tgt = self.norm1(tgt)
-        
-        # Cross-attention
-        tgt2 = self.cross_attention(tgt, memory)
+
+        tgt2 = self.cross_attention(tgt, memory, memory_mask)
         tgt = tgt + self.dropout2(tgt2)
         tgt = self.norm2(tgt)
-        
-        # Feedforward
+
         tgt2 = self.linear2(self.dropout(F.relu(self.linear1(tgt))))
         tgt = tgt + self.dropout3(tgt2)
         tgt = self.norm3(tgt)
@@ -149,27 +147,23 @@ def construct_embedding_vector_from_vocab(command: Commands, parameters: torch.T
     value_tensor = torch.tensor(command.value - 1).cuda()
     one_hot_tensor = F.one_hot(value_tensor, num_classes=num_classes).float().cuda()
 
-    # Ensure one_hot_tensor is 2D by adding a batch dimension if necessary
     if one_hot_tensor.dim() == 1:
-        one_hot_tensor = one_hot_tensor.unsqueeze(0)  # Shape: (1, num_classes)
+        one_hot_tensor = one_hot_tensor.unsqueeze(0)
 
-    # Ensure parameters is 2D and has the correct shape (1, 6)
     if parameters.dim() == 1:
-        parameters = parameters.unsqueeze(0)  # Shape: (1, param_size)
+        parameters = parameters.unsqueeze(0)
 
-    # Ensure parameters has the correct shape
     if parameters.shape[0] != 1:
-        parameters = parameters[:1, :]  # Take only the first row if more than one exists
+        parameters = parameters[:1, :]
     elif parameters.shape[0] == 1 and parameters.shape[1] == 6:
         parameters = parameters
     else:
-        parameters = parameters.view(1, 6)  # Ensure parameters has the shape [1, 6]
+        parameters = parameters.view(1, 6)
 
     embedding_vector = torch.cat((one_hot_tensor, parameters), dim=1).cuda()
     assert embedding_vector.shape == (1, 11), f"Expected shape-> (1, 11), got {embedding_vector.shape}"
-    
-    return embedding_vector
 
+    return embedding_vector
 
 class CommandTransformer(nn.Module):
     def __init__(self, vocab_size, d_model=512, num_layers=6):
@@ -178,52 +172,99 @@ class CommandTransformer(nn.Module):
         self.pos_encoder = PositionalEncoding(d_model).cuda()
         self.transformer = CustomTransformerDecoder(d_model, d_model, d_model, num_layers, 2048).cuda()
         self.output_layer = TransformerOutputLayer(d_model).cuda()
-        self.linear = nn.Linear(11, 512).cuda()  # Linear layer to project each element of tgt to d_model dimensions
+        self.linear = nn.Linear(11, 512).cuda()
 
-    def forward(self, src: torch.Tensor, tgt: torch.Tensor):
-        src_emb = src.unsqueeze(0).cuda()  # Adding batch dimension
-        # print(f"Original tgt shape: {tgt.shape}")
+    def forward(self, src: torch.Tensor, tgt: torch.Tensor, tgt_mask=None, memory_mask=None, tgt_key_padding_mask=None, memory_key_padding_mask=None):
+        src_emb = src.unsqueeze(0).cuda()
 
-        tgt = tgt.view(-1, 11)  # Shape: (batch_size*sequence_length, 11)
-        # tgt = self.linear(tgt)  # Shape: (batch_size*sequence_length, 512)
-        tgt = tgt.view(1, -1, 512)  # Reshape back to (1, N*11, 512)
-        # print(f"Shape after linear layer: {tgt.shape}")
+        tgt = tgt.view(1, -1, 512)
 
         tgt_emb = self.pos_encoder(tgt)
-        # print(f"Shape after positional encoding: {tgt_emb.shape}")
 
-        # Transformer forward pass
-        transformer_output = self.transformer(tgt_emb, src_emb)
-        # print(f"Transformer input shape: {tgt_emb.shape}")
-        # print(f"Transformer output shape: {transformer_output.shape}")
+        transformer_output = self.transformer(tgt_emb, src_emb, tgt_mask=tgt_mask, memory_mask=memory_mask, tgt_key_padding_mask=tgt_key_padding_mask, memory_key_padding_mask=memory_key_padding_mask)
 
-        # Output layer forward pass
         outputs = self.output_layer(transformer_output)
-        # print(f"Output layer shapes: command_probs: {outputs[0].shape}, parameters_probs: {outputs[1].shape}")
 
         return outputs
 
-model = CommandTransformer(vocab_size=VOCAB_SIZE).cuda()
-input_emb = construct_embedding_vector_from_vocab(Commands.START, torch.zeros(6).cuda()).cuda()
-input_emb_proj = model.linear(input_emb).unsqueeze(1)  # Shape: (1, 1, 512)
-input_emb_proj = input_emb_proj.repeat(1, 11, 1)  # Shape: (1, 11, 512)
+# Create a causal mask
+def generate_causal_mask(seq_len):
+    mask = torch.tril(torch.ones((seq_len, seq_len), device='cuda'))
+    return mask
 
-while True:
-    pred = model(pt_cloud_encoded_features, input_emb_proj)
-    command, parameters = select_parameters(*pred)
-    print(command)
-    print(parameters[0])
-    output_emb = construct_embedding_vector_from_vocab(command, parameters).cuda()
 
-    # Apply linear layer to project the new embedding to 512 dimensions
-    output_emb_proj = model.linear(output_emb).unsqueeze(1)  # Shape: (1, 1, 512)
-    output_emb_proj = output_emb_proj.repeat(1, 11, 1)  # Shape: (1, 11, 512)
+if __name__ == "__main__":
 
-    # Concatenate along the sequence dimension
-    input_emb_proj = torch.cat((input_emb_proj, output_emb_proj), dim=1).cuda()
+    # Initialize the model, optimizer, and loss function
+    model = CommandTransformer(vocab_size=VOCAB_SIZE).cuda()
+    optimizer = torch.optim.Adam(model.parameters(), lr=1e-4)
+    criterion = nn.MSELoss()
 
-    if command == Commands.STOP:
-        break
+    processor = SceneScriptProcessor('/home/mseleem/Desktop/3d_SceneScript/0/ase_scene_language.txt')
+    gt_embeddings = processor.process()
+    gt_embeddings_tensor = torch.tensor(gt_embeddings).cuda()
+    positional_encoding = PositionalEncoding(gt_embeddings_tensor.shape[-1]).cuda()
+    gt_embeddings_tensor = positional_encoding(gt_embeddings_tensor)
 
-# print(input_emb_proj)
-print(input_emb_proj.shape)
+    # Define number of epochs
+    num_epochs = 10
+
+    # List to store the loss values
+    losses = []
+
+    # Training loop
+    for epoch in range(num_epochs):
+        input_emb = construct_embedding_vector_from_vocab(Commands.START, torch.zeros(6).cuda()).cuda()
+        input_emb_proj = model.linear(input_emb).unsqueeze(1)
+        input_emb_proj = input_emb_proj.repeat(1, 11, 1)
+
+        model.train()
+        epoch_loss = 0
+        for step in range(gt_embeddings_tensor.size(0)):
+            optimizer.zero_grad()
+
+            # Update the causal mask size based on current input_emb_proj size
+            seq_len = input_emb_proj.size(1)
+            tgt_mask = generate_causal_mask(seq_len).unsqueeze(0)  # Correct the shape here
+
+            pred = model(pt_cloud_encoded_features, input_emb_proj, tgt_mask=tgt_mask)
+            command, parameters = select_parameters(*pred)
+
+            output_emb = construct_embedding_vector_from_vocab(command, parameters).cuda()
+            output_emb_proj = model.linear(output_emb).unsqueeze(1)
+            output_emb_proj = output_emb_proj.repeat(1, 11, 1)
+
+            input_emb_proj = torch.cat((input_emb_proj, output_emb_proj), dim=1).cuda()
+
+            # Compute loss with the revealed ground truth
+            loss = criterion(input_emb_proj[:, -1, :], gt_embeddings_tensor[step].unsqueeze(0))
+
+            loss.backward()
+            optimizer.step()
+
+            epoch_loss += loss.item()
+
+            if command == Commands.STOP:
+                break
+
+        avg_loss = epoch_loss / gt_embeddings_tensor.size(0)
+        losses.append(avg_loss)
+        print(f"Epoch [{epoch + 1}/{num_epochs}], Loss: {avg_loss}")
+
+    # Debug: print losses list to ensure it has values
+    print("Losses:", losses)
+
+    # Check if losses is empty
+    if len(losses) == 0:
+        print("Error: Losses list is empty, cannot plot.")
+    else:
+        # Plot the loss
+        plt.figure()  # Ensure a new figure is created
+        plt.plot(losses)
+        plt.xlabel('Epoch')
+        plt.ylabel('Loss')
+        plt.title('Training Loss Over Epochs')
+        plt.grid(True)  # Add grid for better readability
+        plt.show()
+
+    print("Training complete.")
