+++ scripts/pt_cloud_softmax.py
@@ -1,13 +1,11 @@
 import torch
-from tests.point_cloud_sinusoidal_pt_cloud_no_vit import PointCloudTransformerLayer
+from three_d_scene_script.point_cloud_sinusoidal_pt_cloud_no_vit import PointCloudTransformerLayer
 import torch.nn as nn
 import torch.nn.functional as F
 import math
 from enum import Enum
-from test_gt import SceneScriptProcessor
-import matplotlib.pyplot as plt
 
-# Initialize the models and other components
+# Encoder model setup
 encoder_model = PointCloudTransformerLayer().cuda()
 pt_cloud_path = "/home/mseleem/Desktop/3d_model_pt/0/semidense_points.csv.gz"
 points, dist_std = encoder_model.read_points_file(pt_cloud_path)
@@ -34,22 +32,25 @@ class Commands(Enum):
         if value == cls.MAKE_DOOR.value:
             return cls.MAKE_DOOR
 
+# Output layer for commands and parameters
 class TransformerOutputLayer(nn.Module):
     def __init__(self, transformer_dim):
         super(TransformerOutputLayer, self).__init__()
-        self.command_layer = nn.Linear(transformer_dim, 5)
-        self.param_layer = nn.Linear(transformer_dim, 6)
+        self.command_layer = nn.Linear(transformer_dim, 5)  # Output 5 command logits (START included)
+        self.param_layer = nn.Linear(transformer_dim, 6)  # Output 6 params for wall, doors, windows 
 
     def forward(self, x):
-        command_logits = self.command_layer(x)
-        command_logits[..., 0] = -float('inf')
-        command_probs = F.softmax(command_logits, dim=-1)
-        parameter_logits = self.param_layer(x)
-        parameters_probs = torch.tanh(parameter_logits)
+        command_logits = self.command_layer(x)  # Shape: (batch_size, sequence_length, 5)
+        command_probs = F.softmax(command_logits, dim=-1)  # Shape: (batch_size, sequence_length, 5)
+        parameter_logits = self.param_layer(x)  # Shape: (batch_size, sequence_length, 6)
+        parameters_probs = F.softmax(parameter_logits, dim=-1)  # Shape: (batch_size, sequence_length, 6)
         return command_probs, parameters_probs
 
+# Function to select parameters based on command probabilities
 def select_parameters(command_probs, parameters_probs):
-    command_indx = command_probs[0, 0].argmax(dim=-1).item() + 1
+    command_indx = command_probs[0, 0].argmax(dim=-1).item() + 1 
+    # command_indx = command_probs.argmax(dim=-1) + 2
+    # command_indx = command_indx.numpy().item()
     if command_indx == Commands.STOP.value:
         parameters = torch.zeros(6).cuda()
     elif command_indx in [Commands.MAKE_WALL.value, Commands.MAKE_DOOR.value, Commands.MAKE_WINDOW.value]:
@@ -83,20 +84,18 @@ class CrossAttention(nn.Module):
     def __init__(self, d_model, d_out_kq, d_out_v):
         super(CrossAttention, self).__init__()
         self.d_out_kq = d_out_kq
-        self.W_query = nn.Linear(d_model, d_out_kq)
-        self.W_key = nn.Linear(d_model, d_out_kq)
-        self.W_value = nn.Linear(d_model, d_out_v)
-
-    def forward(self, x_1, x_2, mask=None):
-        queries_1 = self.W_query(x_1)
-        keys_2 = self.W_key(x_2)
-        values_2 = self.W_value(x_2)
-
+        self.W_query = nn.Parameter(torch.rand(d_model, d_out_kq))
+        self.W_key = nn.Parameter(torch.rand(d_model, d_out_kq))
+        self.W_value = nn.Parameter(torch.rand(d_model, d_out_v))
+    
+    def forward(self, x_1, x_2):
+        queries_1 = x_1.matmul(self.W_query)
+        keys_2 = x_2.matmul(self.W_key)
+        values_2 = x_2.matmul(self.W_value)
+        
         attn_scores = queries_1.matmul(keys_2.transpose(-2, -1))
-        if mask is not None:
-            attn_scores = attn_scores.masked_fill(mask == 0, float('-inf'))
         attn_weights = torch.softmax(attn_scores / self.d_out_kq**0.5, dim=-1)
-
+        
         context_vec = attn_weights.matmul(values_2)
         return context_vec
 
@@ -116,14 +115,17 @@ class CustomTransformerDecoderLayer(nn.Module):
         self.dropout3 = nn.Dropout(dropout)
 
     def forward(self, tgt, memory, tgt_mask=None, memory_mask=None, tgt_key_padding_mask=None, memory_key_padding_mask=None):
+        # Self-attention
         tgt2, _ = self.self_attn(tgt, tgt, tgt, attn_mask=tgt_mask, key_padding_mask=tgt_key_padding_mask)
         tgt = tgt + self.dropout1(tgt2)
         tgt = self.norm1(tgt)
-
-        tgt2 = self.cross_attention(tgt, memory, memory_mask)
+        
+        # Cross-attention
+        tgt2 = self.cross_attention(tgt, memory)
         tgt = tgt + self.dropout2(tgt2)
         tgt = self.norm2(tgt)
-
+        
+        # Feedforward
         tgt2 = self.linear2(self.dropout(F.relu(self.linear1(tgt))))
         tgt = tgt + self.dropout3(tgt2)
         tgt = self.norm3(tgt)
@@ -138,32 +140,45 @@ class CustomTransformerDecoder(nn.Module):
 
     def forward(self, tgt, memory, tgt_mask=None, memory_mask=None, tgt_key_padding_mask=None, memory_key_padding_mask=None):
         for layer in self.layers:
+            # print(f"Layer input shapes - tgt: {tgt.shape}, memory: {memory.shape}")
             tgt = layer(tgt, memory, tgt_mask=tgt_mask, memory_mask=memory_mask,
                         tgt_key_padding_mask=tgt_key_padding_mask, memory_key_padding_mask=memory_key_padding_mask)
+            # print(f"After layer tgt shape: {tgt.shape}")
         return tgt
 
 def construct_embedding_vector_from_vocab(command: Commands, parameters: torch.Tensor):
     num_classes = len(Commands)
+
+    # Convert the integer value to a tensor
     value_tensor = torch.tensor(command.value - 1).cuda()
+
+    # Create the one-hot tensor
     one_hot_tensor = F.one_hot(value_tensor, num_classes=num_classes).float().cuda()
 
+    # Ensure one_hot_tensor is 2D by adding a batch dimension if necessary
     if one_hot_tensor.dim() == 1:
-        one_hot_tensor = one_hot_tensor.unsqueeze(0)
+        one_hot_tensor = one_hot_tensor.unsqueeze(0)  # Shape: (1, num_classes)
 
+    # Ensure parameters is 2D and has the correct shape (1, 6)
     if parameters.dim() == 1:
-        parameters = parameters.unsqueeze(0)
+        parameters = parameters.unsqueeze(0)  # Shape: (1, param_size)
 
+    # Ensure parameters has the correct shape
     if parameters.shape[0] != 1:
-        parameters = parameters[:1, :]
+        parameters = parameters[:1, :]  # Take only the first row if more than one exists
     elif parameters.shape[0] == 1 and parameters.shape[1] == 6:
         parameters = parameters
     else:
-        parameters = parameters.view(1, 6)
+        parameters = parameters.view(1, 6)  # Ensure parameters has the shape [1, 6]
 
+    # Concatenate the one-hot tensor and parameters along the feature dimension
     embedding_vector = torch.cat((one_hot_tensor, parameters), dim=1).cuda()
-    assert embedding_vector.shape == (1, 11), f"Expected shape-> (1, 11), got {embedding_vector.shape}"
-
+    
+    # Ensure the output shape is [1, 11]
+    assert embedding_vector.shape == (1, 11), f"Expected shape (1, 11), but got {embedding_vector.shape}"
+    
     return embedding_vector
+
 
 class CommandTransformer(nn.Module):
     def __init__(self, vocab_size, d_model=512, num_layers=6):
@@ -172,96 +187,62 @@ class CommandTransformer(nn.Module):
         self.pos_encoder = PositionalEncoding(d_model).cuda()
         self.transformer = CustomTransformerDecoder(d_model, d_model, d_model, num_layers, 2048).cuda()
         self.output_layer = TransformerOutputLayer(d_model).cuda()
-        self.linear = nn.Linear(11, 512).cuda()
+        self.linear = nn.Linear(vocab_size, d_model).cuda()  # Linear layer to project each element of tgt to d_model dimensions
 
-    def forward(self, src: torch.Tensor, tgt: torch.Tensor, tgt_mask=None, memory_mask=None, tgt_key_padding_mask=None, memory_key_padding_mask=None):
-        src_emb = src.unsqueeze(0).cuda()
+    def forward(self, src: torch.Tensor, tgt: torch.Tensor):
+        src_emb = src.unsqueeze(0).cuda()  # Add batch dimension
+        tgt = tgt.unsqueeze(0).cuda()  # Add batch dimension
 
-        tgt = tgt.view(1, -1, 512)
+        # print(f"Original tgt shape: {tgt.shape}")
 
+        # Apply the linear layer
+        tgt = self.linear(tgt)
+        # print(f"Shape after linear layer: {tgt.shape}")
+
+        # Ensure tgt has the correct shape (batch_size, seq_len, d_model)
+        tgt = tgt.view(tgt.size(0), -1, 512)
+        # print(f"Shape after reshaping: {tgt.shape}")
+
+        # Apply positional encoding
         tgt_emb = self.pos_encoder(tgt)
+        # print(f"Positional encoded tgt shape: {tgt_emb.shape}")
 
-        transformer_output = self.transformer(tgt_emb, src_emb, tgt_mask=tgt_mask, memory_mask=memory_mask, tgt_key_padding_mask=tgt_key_padding_mask, memory_key_padding_mask=memory_key_padding_mask)
+        # Check dimensions before transformer
+        assert tgt_emb.dim() == 3, f"Expected 3D tensor for tgt_emb, got {tgt_emb.dim()}D"
+        assert src_emb.dim() == 3, f"Expected 3D tensor for src_emb, got {src_emb.dim()}D"
 
+        # Transformer forward pass
+        transformer_output = self.transformer(tgt_emb, src_emb)
+        # print(f"Transformer output shape: {transformer_output.shape}")
+
+        # Output layer forward pass
         outputs = self.output_layer(transformer_output)
+        # print(f"Output layer shape: {outputs[0].shape}, {outputs[1].shape}")
 
         return outputs
 
-# Create a causal mask
-def generate_causal_mask(seq_len):
-    mask = torch.tril(torch.ones((seq_len, seq_len), device='cuda'))
-    return mask
 
-# Initialize the model, optimizer, and loss function
-model = CommandTransformer(vocab_size=VOCAB_SIZE).cuda()
-optimizer = torch.optim.Adam(model.parameters(), lr=1e-4)
-criterion = nn.MSELoss()
+if __name__ == "__main__":
 
-processor = SceneScriptProcessor('/home/mseleem/Desktop/3d_SceneScript/0/ase_scene_language.txt')
-gt_embeddings = processor.process()
-gt_embeddings_tensor = torch.tensor(gt_embeddings).cuda()
-positional_encoding = PositionalEncoding(gt_embeddings_tensor.shape[-1]).cuda()
-gt_embeddings_tensor = positional_encoding(gt_embeddings_tensor)
-
-# Define number of epochs
-num_epochs = 10
-
-# List to store the loss values
-losses = []
-
-# Training loop
-for epoch in range(num_epochs):
+    # Initialize the model
+    model = CommandTransformer(vocab_size=VOCAB_SIZE).cuda()
     input_emb = construct_embedding_vector_from_vocab(Commands.START, torch.zeros(6).cuda()).cuda()
-    input_emb_proj = model.linear(input_emb).unsqueeze(1)
-    input_emb_proj = input_emb_proj.repeat(1, 11, 1)
 
-    model.train()
-    epoch_loss = 0
-    for step in range(gt_embeddings_tensor.size(0)):
-        optimizer.zero_grad()
-
-        # Update the causal mask size based on current input_emb_proj size
-        seq_len = input_emb_proj.size(1)
-        tgt_mask = generate_causal_mask(seq_len).unsqueeze(0)  # Correct the shape here
-
-        pred = model(pt_cloud_encoded_features, input_emb_proj, tgt_mask=tgt_mask)
+    while True:
+        pred = model(pt_cloud_encoded_features, input_emb)
         command, parameters = select_parameters(*pred)
-
         output_emb = construct_embedding_vector_from_vocab(command, parameters).cuda()
-        output_emb_proj = model.linear(output_emb).unsqueeze(1)
-        output_emb_proj = output_emb_proj.repeat(1, 11, 1)
 
-        input_emb_proj = torch.cat((input_emb_proj, output_emb_proj), dim=1).cuda()
+        # Debug the shapes before concatenation
+        # print(f"input_emb shape before concatenation: {input_emb.shape}")
+        # print(f"output_emb shape before concatenation: {output_emb.shape}")
 
-        # Compute loss with the revealed ground truth
-        loss = criterion(input_emb_proj[:, -1, :], gt_embeddings_tensor[step].unsqueeze(0))
-
-        loss.backward()
-        optimizer.step()
-
-        epoch_loss += loss.item()
+        # Ensure dimensions align for concatenation
+        input_emb = torch.cat((input_emb, output_emb), dim=0).cuda()  # Concatenate along the sequence dimension
 
         if command == Commands.STOP:
             break
 
-    avg_loss = epoch_loss / gt_embeddings_tensor.size(0)
-    losses.append(avg_loss)
-    print(f"Epoch [{epoch + 1}/{num_epochs}], Loss: {avg_loss}")
+    print(input_emb)
+    print(input_emb.shape)
 
-# Debug: print losses list to ensure it has values
-print("Losses:", losses)
-
-# Check if losses is empty
-if len(losses) == 0:
-    print("Error: Losses list is empty, cannot plot.")
-else:
-    # Plot the loss
-    plt.figure()  # Ensure a new figure is created
-    plt.plot(losses)
-    plt.xlabel('Epoch')
-    plt.ylabel('Loss')
-    plt.title('Training Loss Over Epochs')
-    plt.grid(True)  # Add grid for better readability
-    plt.show()
-
-print("Training complete.")
