+++ src/three_d_scene_script/cont_emb.py
