+++ src/three_d_scene_script/point_cloud_encoder_cat_active_sites.py
