+++ src/three_d_scene_script/point_cloud_sinusoidal_cont_emb.py
