+++ src/three_d_scene_script/point_cloud_sinusoidal_pt_cloud_no_vit.py
